server/requirements.txt
asgiref==3.5.2
daphne==3.0.2
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary==2.9.1
gunicorn==20.1.0
requests==2.20.1
django-postgres-copy==2.7.1
python-dateutil==2.8.2
redis==4.3.4
celery==5.2.7
channels==3.0.5
channels-redis==3.4.1
pycoingecko==2.2.0
cffi==1.15.1